refactor(server): update billing module types
