--- packages/amplication-server/src/core/billing/BillingFeature.ts
@@ -1,11 +1,0 @@
-export enum BillingFeature {
-  Services = "feature-services",
-  TeamMembers = "feature-team-members",
-  EntitiesPerService = "feature-entities-per-service",
-  ServicesWithManyEntities = "feature-services-wth-many-entities",
-  CodeGenerationBuilds = "feature-code-generation-builds",
-  CodePushToGit = "feature-code-push-to-git",
-  IgnoreValidationCodeGeneration = "feature-ignore-validation-code-generation",
-  HideNotifications = "feature-ignore-validation-code-generation-hide-notifications",
-  ServicesAboveEntitiesPerServiceLimit = "feature-services-above-entities-per-service-limit",
-}
