--- packages/amplication-server/src/core/billing/BillingPlan.ts
@@ -1,5 +1,0 @@
-export enum BillingPlan {
-  Free = "plan-amplication-free",
-  Pro = "plan-amplication-pro",
-  Enterprise = "plan-amplication-enterprise",
-}
